requirements.txt
bitstring==3.1.5
BitTorrent-bencode==5.0.8.1
certifi==2018.11.29
chardet==3.0.4
idna==2.7
requests==2.20.1
sha1==1.2
urllib3==1.24.1
